Changes: - Auto resize text box to show all text for self-explanation items. - Change AudioMainSettingsDialogBox resize mode to auto fit content size.

+++ PowerPointLabs/PowerPointLabs/ELearningLab/Views/AudioSettingsViews/AzureVoiceLoginPage.xaml.cs
@@ -82,7 +82,7 @@ namespace PowerPointLabs.ELearningLab.Views
                     dialog.SetDialogWindowHeight(AudioSettingService.AudioMainSettingsPageHeight);
                     break;
                 case AudioSettingsPage.AudioPreviewPage:
-                    dialog.SetDialogWindowHeight(AudioSettingService.AudioPreviewPageHeight);
+                    dialog.SetDialogWindowHeight((int)AudioSettingService.AudioPreviewPageHeight);
                     break;
             }
             dialog.SetCurrentPage(previousPage);
